Ids serializers y Score con floats

+++ restemplate/models.py
@@ -5,7 +5,7 @@ from django.core.validators import MinValueValidator, MaxValueValidator
 class Rating(models.Model):
     song = models.IntegerField()
     user = models.IntegerField()
-    score = models.PositiveSmallIntegerField(validators = [MinValueValidator(0), MaxValueValidator(10)])
+    score = models.FloatField(validators = [MinValueValidator(0), MaxValueValidator(10)])
 
     class Meta:
         unique_together = ('song', 'user')
